Fix #153: fix: raise agentready score from Silver (67.8) to Gold (75+)
Address all 7 failing agentready attributes:

- CI Quality Gates: add explicit "Type Check" and "Static Analysis" steps
- Single-File Verification: add per-file verification commands to CLAUDE.md
- Pattern References: add "Common Patterns" section to CLAUDE.md
- Code Smell Elimination: enable errcheck, gosimple, govet, ineffassign,
  staticcheck, unused linters in .golangci.yml
- Concise Documentation: reduce README heading density from 10.8 to 4.2
  per 100 lines (target 3-5), consolidate sections
- Design Intent: add design decisions, invariants, and constraints to
  specs/architecture.md and specs/error-handling.md
- File Size Limits: split loop.go (1899 lines) into 7 files by concern,
  all under 1000 lines
