+++ pkg/agent/loop.go
@@ -113,7 +113,6 @@ func runSequential[T any](ctx context.Context, items []T, fn func(context.Contex
 	}
 }
 
-// NewAgent creates a new Agent with all dependencies wired.
 // Task structs for parallel execution
 
 type newIssueTask struct {
@@ -146,6 +145,7 @@ type conflictTask struct {
 	rebaseStderr string
 }
 
+// NewAgent creates a new Agent with all dependencies wired.
 func NewAgent(gh GitHubClient, runner CommandRunner, worktrees WorktreeManager, state *State, cfg Config, logger *slog.Logger, codeAgent CodeAgent) *Agent {
 	return &Agent{
 		gh:        gh,
@@ -206,1299 +206,6 @@ func (a *Agent) EmitPollCycleEnd() {
 		State:  "idle",
 		Action: "Poll cycle completed",
 	})
-}
-
-// ProcessNewIssues finds labeled issues and spawns Claude to implement fixes.
-func (a *Agent) ProcessNewIssues(ctx context.Context) {
-	if a.cfg.Label == "" {
-		return
-	}
-
-	a.emit(Event{
-		Type:   EventActionStarted,
-		Worker: a.workerName(),
-		State:  "working",
-		Action: "Scanning for new issues",
-	})
-	defer a.emit(Event{
-		Type:   EventActionCompleted,
-		Worker: a.workerName(),
-		State:  "idle",
-		Action: "Issue scanning complete",
-	})
-
-	issues, err := a.gh.ListLabeledIssues(ctx, a.cfg.Owner, a.cfg.Repo, a.cfg.Label)
-	if err != nil {
-		a.logger.Error("failed to list issues", "error", err)
-		a.emit(Event{Type: EventError, Worker: a.workerName(), State: "error", Error: "failed to list issues: " + err.Error()})
-		return
-	}
-
-	// Sequential phase: filter issues, create worktrees, insert into state
-	var tasks []newIssueTask
-
-	for _, issue := range issues {
-		issueKey := IssueKey(a.cfg.Owner, a.cfg.Repo, issue.Number)
-		if work, exists := a.state.ActiveIssues[issueKey]; exists {
-			// Re-check for PR if we lost track of it
-			if work.PRNumber == 0 && work.Status == StatusImplementing {
-				prs, err := a.gh.ListPRsByHead(ctx, a.cfg.Owner, a.cfg.Repo, a.cfg.GitHubHeadOwner, work.BranchName)
-				if err == nil {
-					openPR, _ := classifyPRs(prs)
-					if openPR != nil {
-						work.PRNumber = openPR.Number
-						work.Status = StatusPROpen
-						a.logger.Info("found PR for tracked issue", "issue", issue.Number, "pr", work.PRNumber)
-					}
-				}
-			}
-			a.logger.Debug("skipping already tracked issue", "issue", issue.Number)
-			continue
-		}
-
-		if a.cfg.OnlyAssigned && !issueAssignedTo(issue, a.cfg.GitHubUser) {
-			a.logger.Debug("skipping issue not assigned to agent user", "issue", issue.Number, "user", a.cfg.GitHubUser)
-			continue
-		}
-
-		a.logger.Info("processing new issue", "issue", issue.Number, "title", issue.Title)
-
-		branchName := issueBranchName(issue.Number)
-
-		// Check if a PR already exists for this issue (open, closed, or merged)
-		prs, err := a.gh.ListPRsByHead(ctx, a.cfg.Owner, a.cfg.Repo, a.cfg.GitHubHeadOwner, branchName)
-		if err == nil && len(prs) > 0 {
-			// Find the first open PR, or check if any was merged
-			openPR, hasMerged := classifyPRs(prs)
-			if openPR != nil {
-				a.logger.Info("PR already exists for issue", "issue", issue.Number, "pr", openPR.Number)
-				a.state.ActiveIssues[issueKey] = &IssueWork{
-					IssueNumber: issue.Number,
-					IssueTitle:  issue.Title,
-					BranchName:  branchName,
-					PRNumber:    openPR.Number,
-					Status:      StatusPROpen,
-					CreatedAt:   time.Now(),
-				}
-				continue
-			} else if hasMerged {
-				// PR was merged — skip to avoid reprocessing a completed issue
-				a.logger.Info("skipping issue with merged PR", "issue", issue.Number, "pr", prs[0].Number)
-				continue
-			}
-			// PR was closed (rejected) — fall through to allow retry
-		}
-
-		// Check if any open PR already references this issue (e.g., created by a human)
-		if linked, err := a.gh.HasLinkedPR(ctx, a.cfg.Owner, a.cfg.Repo, issue.Number); err != nil {
-			a.logger.Warn("failed to check for linked PRs", "issue", issue.Number, "error", err)
-		} else if linked {
-			a.logger.Info("skipping issue with existing linked PR", "issue", issue.Number)
-			continue
-		}
-
-		// Only post in-progress comment if we haven't already
-		if !a.hasExistingBotComment(ctx, issue.Number, "working on this issue") {
-			if err := a.gh.AssignIssue(ctx, a.cfg.Owner, a.cfg.Repo, issue.Number, a.cfg.GitHubUser); err != nil {
-				a.logger.Warn("failed to assign issue", "issue", issue.Number, "error", err)
-			}
-			if !a.ShouldSkipComment("issue-in-progress") {
-				if err := a.gh.AddIssueComment(ctx, a.cfg.Owner, a.cfg.Repo, issue.Number,
-					fmt.Sprintf("Oompa is working on this issue. A PR will be created shortly.\n\n%s", a.botComment())); err != nil {
-					a.logger.Warn("failed to add in-progress comment", "issue", issue.Number, "error", err)
-				}
-			}
-		}
-
-		if err := a.worktrees.EnsureRepoCloned(ctx); err != nil {
-			a.logger.Error("failed to ensure repo cloned", "error", err)
-			return
-		}
-
-		worktreePath, err := a.worktrees.CreateWorktree(ctx, branchName)
-		if err != nil {
-			a.logger.Error("failed to create worktree", "issue", issue.Number, "error", err)
-			continue
-		}
-
-		work := &IssueWork{
-			IssueNumber:  issue.Number,
-			IssueTitle:   issue.Title,
-			WorktreePath: worktreePath,
-			BranchName:   branchName,
-			Status:       StatusImplementing,
-			CreatedAt:    time.Now(),
-		}
-
-		// Insert into state map before parallel phase
-		a.state.ActiveIssues[issueKey] = work
-
-		tasks = append(tasks, newIssueTask{
-			issue:        issue,
-			branchName:   branchName,
-			worktreePath: worktreePath,
-			work:         work,
-		})
-	}
-
-	// Parallel phase: run Claude, push, create PR
-	runSequential(ctx, tasks, func(ctx context.Context, task newIssueTask) {
-		prompt := buildImplementationPrompt(task.issue, a.cfg.SignedOffBy)
-		_, err := a.codeAgent.Run(ctx, a.runner, task.worktreePath, prompt, a.logger, false)
-		if err != nil {
-			a.logger.Error("agent failed", "issue", task.issue.Number, "error", err)
-			a.markIssueFailed(ctx, task.issue.Number, task.work)
-			return
-		}
-
-		// Check if Claude produced any commits
-		logOut, _, _ := a.runner.Run(ctx, task.worktreePath, "git", "log", a.originDefaultBranch()+"..HEAD", "--oneline")
-		if strings.TrimSpace(string(logOut)) == "" {
-			a.logger.Warn("claude finished but produced no commits", "issue", task.issue.Number)
-			a.markIssueFailed(ctx, task.issue.Number, task.work)
-			return
-		}
-
-		// Squash all commits into a single commit before pushing
-		if err := a.gitSquashCommits(ctx, task.worktreePath, task.issue.Number, task.issue.Title); err != nil {
-			a.logger.Error("failed to squash commits", "issue", task.issue.Number, "error", err)
-			a.markIssueFailed(ctx, task.issue.Number, task.work)
-			return
-		}
-
-		// Push the branch (force push because squashing rewrites history)
-		if err := a.gitPush(ctx, task.worktreePath, true); err != nil {
-			a.logger.Error("failed to push branch", "issue", task.issue.Number, "error", err)
-			a.markIssueFailed(ctx, task.issue.Number, task.work)
-			return
-		}
-
-		// Create PR
-		prHead := task.branchName
-		if a.cfg.GitHubHeadOwner != "" && a.cfg.GitHubHeadOwner != a.cfg.Owner {
-			prHead = a.cfg.GitHubHeadOwner + ":" + task.branchName
-		}
-		prTitle := task.issue.Title
-		prBody := a.buildPRBody(ctx, task.worktreePath, task.issue.Number)
-		prNumber, err := a.gh.CreatePR(ctx, a.cfg.Owner, a.cfg.Repo, prTitle, prBody, prHead, a.defaultBranch())
-		if err != nil {
-			a.logger.Error("failed to create PR", "issue", task.issue.Number, "error", err)
-			// Clean up the remote branch to avoid orphaned branches
-			a.deleteRemoteBranch(ctx, task.worktreePath, task.branchName)
-			a.markIssueFailed(ctx, task.issue.Number, task.work)
-			return
-		}
-
-		task.work.PRNumber = prNumber
-		task.work.Status = StatusPROpen
-		a.logger.Info("created PR", "issue", task.issue.Number, "pr", prNumber)
-
-		_ = a.gh.UnassignIssue(ctx, a.cfg.Owner, a.cfg.Repo, task.issue.Number, a.cfg.GitHubUser)
-	})
-}
-
-// ProcessReviewComments checks for new review comments and review bodies, then runs Claude to address them.
-func (a *Agent) ProcessReviewComments(ctx context.Context) {
-	a.emit(Event{
-		Type:   EventActionStarted,
-		Worker: a.workerName(),
-		State:  "reviewing",
-		Action: "Checking for review comments",
-	})
-	defer a.emit(Event{
-		Type:   EventActionCompleted,
-		Worker: a.workerName(),
-		State:  "idle",
-		Action: "Review check complete",
-	})
-	// Sequential phase: filter comments, prepare worktrees, add reactions, build tasks
-	var tasks []reviewTask
-
-	for _, work := range a.state.ActiveIssues {
-		if work.PRNumber == 0 || work.Status != StatusPROpen {
-			continue
-		}
-
-		comments, err := a.gh.GetPRReviewComments(ctx, a.cfg.Owner, a.cfg.Repo, work.PRNumber, work.LastCommentID)
-		if err != nil {
-			a.logger.Error("failed to get PR comments", "pr", work.PRNumber, "error", err)
-			continue
-		}
-
-		// Track the max comment ID across ALL fetched comments (including filtered ones)
-		// so the cursor advances past bot-posted/already-replied comments that would
-		// otherwise be re-fetched and re-filtered on every poll cycle.
-		var maxCommentID int64
-		for _, c := range comments {
-			if c.ID > maxCommentID {
-				maxCommentID = c.ID
-			}
-		}
-
-		// Build a set of comment IDs that have a reply from our user
-		repliedTo := make(map[int64]bool)
-		for _, c := range comments {
-			if c.InReplyToID != 0 && c.User == a.cfg.GitHubUser {
-				repliedTo[c.InReplyToID] = true
-			}
-		}
-
-		// Filter comments: skip replies, skip bot-posted, only whitelisted reviewers, skip already-processed
-		var humanComments []ReviewComment
-		for _, c := range comments {
-			if c.InReplyToID != 0 {
-				continue
-			}
-			if strings.Contains(c.Body, botMarker) {
-				continue
-			}
-			if !a.isAllowedReviewer(c.User) {
-				continue
-			}
-			if repliedTo[c.ID] {
-				if already, err := a.gh.HasPRCommentReaction(ctx, a.cfg.Owner, a.cfg.Repo, c.ID, "eyes", a.cfg.GitHubUser); err == nil && already {
-					continue
-				}
-			}
-			humanComments = append(humanComments, c)
-		}
-
-		// Fetch PR review bodies
-		reviews, err := a.gh.GetPRReviews(ctx, a.cfg.Owner, a.cfg.Repo, work.PRNumber, work.LastReviewID)
-		if err != nil {
-			a.logger.Warn("failed to get PR reviews", "pr", work.PRNumber, "error", err)
-		}
-
-		// Track the max review ID across ALL fetched reviews (including filtered ones).
-		var maxReviewID int64
-		for _, r := range reviews {
-			if r.ID > maxReviewID {
-				maxReviewID = r.ID
-			}
-		}
-
-		// Filter reviews
-		var humanReviews []PRReview
-		if len(reviews) > 0 {
-			headCommitDate, _ := a.gh.GetPRHeadCommitDate(ctx, a.cfg.Owner, a.cfg.Repo, work.PRNumber)
-
-			for _, r := range reviews {
-				if strings.Contains(r.Body, botMarker) {
-					continue
-				}
-				if !a.isAllowedReviewer(r.User) {
-					continue
-				}
-				if !headCommitDate.IsZero() && r.SubmittedAt.Before(headCommitDate) {
-					continue
-				}
-				humanReviews = append(humanReviews, r)
-			}
-		}
-
-		if len(humanComments) == 0 && len(humanReviews) == 0 {
-			// No actionable comments/reviews, but still advance cursors past
-			// filtered items to avoid re-fetching them on every poll cycle.
-			if maxCommentID > work.LastCommentID {
-				work.LastCommentID = maxCommentID
-			}
-			if maxReviewID > work.LastReviewID {
-				work.LastReviewID = maxReviewID
-			}
-			continue
-		}
-
-		a.logger.Info("addressing review feedback", "pr", work.PRNumber, "comments", len(humanComments), "reviews", len(humanReviews))
-
-		if err := a.ensureWorktreeReady(ctx, work); err != nil {
-			a.logger.Error("failed to prepare worktree", "pr", work.PRNumber, "error", err)
-			continue
-		}
-
-		// React with eyes to signal we're processing
-		for _, c := range humanComments {
-			if err := a.gh.AddPRCommentReaction(ctx, a.cfg.Owner, a.cfg.Repo, c.ID, "eyes"); err != nil {
-				a.logger.Warn("failed to add reaction", "comment", c.ID, "error", err)
-			}
-		}
-
-		tasks = append(tasks, reviewTask{
-			work:          work,
-			humanComments: humanComments,
-			humanReviews:  humanReviews,
-			maxCommentID:  maxCommentID,
-			maxReviewID:   maxReviewID,
-		})
-	}
-
-	// Parallel phase: run agent to address review feedback, then push changes
-	runSequential(ctx, tasks, func(ctx context.Context, task reviewTask) {
-		a.emit(Event{
-			Type:      EventAgentInvocation,
-			Worker:    a.workerName(),
-			State:     "working",
-			Action:    "Addressing review feedback",
-			PRNumbers: []int{task.work.PRNumber},
-		})
-		agentStart := time.Now()
-
-		// Capture local HEAD before agent runs so we can detect if it committed directly
-		headBefore, _, err := a.runner.Run(ctx, task.work.WorktreePath, "git", "rev-parse", "HEAD")
-		if err != nil {
-			a.logger.Warn("failed to get HEAD before agent", "pr", task.work.PRNumber, "error", err)
-		}
-		headSHABefore := strings.TrimSpace(string(headBefore))
-
-		prompt := buildReviewResponsePrompt(*task.work, task.humanComments, task.humanReviews, a.cfg.Owner, a.cfg.Repo)
-		_, err = a.codeAgent.Run(ctx, a.runner, task.work.WorktreePath, prompt, a.logger, true)
-		if err != nil {
-			a.logger.Error("agent failed to address review", "pr", task.work.PRNumber, "error", err)
-			a.emit(Event{
-				Type:      EventError,
-				Worker:    a.workerName(),
-				State:     "error",
-				Action:    "Agent failed to address review",
-				PRNumbers: []int{task.work.PRNumber},
-				Duration:  time.Since(agentStart).Seconds(),
-				Error:     err.Error(),
-			})
-			return
-		}
-		a.emit(Event{
-			Type:      EventAgentCompleted,
-			Worker:    a.workerName(),
-			State:     "idle",
-			Action:    "Review feedback addressed",
-			PRNumbers: []int{task.work.PRNumber},
-			Duration:  time.Since(agentStart).Seconds(),
-		})
-
-		// Push if agent made changes (uncommitted or committed directly).
-		// Track whether push succeeded so we can gate cursor advancement:
-		// if changes were detected but push failed, don't advance the cursor
-		// so the comments are retried on the next poll cycle.
-		pushed, changeDetected := false, false
-		hasFixups := a.hasFixupCommits(ctx, task.work.WorktreePath)
-		hasUncommitted := a.hasUncommittedChanges(ctx, task.work.WorktreePath)
-
-		switch {
-		case hasFixups:
-			// Agent created fixup commits — autosquash them into their targets
-			changeDetected = true
-			if err := a.gitAutosquashRebase(ctx, task.work.WorktreePath); err != nil {
-				a.logger.Error("failed to autosquash fixup commits", "pr", task.work.PRNumber, "error", err)
-			} else if err := a.gitPush(ctx, task.work.WorktreePath, true); err != nil {
-				a.logger.Error("failed to push", "pr", task.work.PRNumber, "error", err)
-			} else {
-				pushed = true
-			}
-		case hasUncommitted:
-			changeDetected = true
-			if err := a.gitAmendAll(ctx, task.work.WorktreePath); err != nil {
-				a.logger.Error("failed to amend commit", "pr", task.work.PRNumber, "error", err)
-			} else if err := a.gitPush(ctx, task.work.WorktreePath, true); err != nil {
-				a.logger.Error("failed to push", "pr", task.work.PRNumber, "error", err)
-			} else {
-				pushed = true
-			}
-		default:
-			headAfter, _, revErr := a.runner.Run(ctx, task.work.WorktreePath, "git", "rev-parse", "HEAD")
-			headSHAAfter := strings.TrimSpace(string(headAfter))
-			if revErr == nil && headSHABefore != "" && headSHAAfter != headSHABefore {
-				changeDetected = true
-				// Agent committed directly — squash new commits into the original HEAD
-				// to avoid polluting commit history with "Address PR review feedback" commits.
-				a.logger.Info("agent committed directly, squashing into original HEAD", "pr", task.work.PRNumber)
-				if err := a.gitSquashInto(ctx, task.work.WorktreePath, headSHABefore); err != nil {
-					a.logger.Error("failed to squash agent commits, skipping push to avoid unsquashed commits", "pr", task.work.PRNumber, "error", err)
-					// Restore HEAD so the worktree is usable on the next attempt
-					if _, _, resetErr := a.runner.Run(ctx, task.work.WorktreePath, "git", "reset", "--hard", headSHAAfter); resetErr != nil {
-						a.logger.Error("failed to restore HEAD after squash failure", "pr", task.work.PRNumber, "error", resetErr)
-					}
-				} else if err := a.gitPush(ctx, task.work.WorktreePath, true); err != nil {
-					a.logger.Error("failed to push", "pr", task.work.PRNumber, "error", err)
-				} else {
-					pushed = true
-				}
-			}
-		}
-
-		// Advance cursors based on ALL fetched comment/review IDs (not just
-		// the human-filtered subset) to avoid re-fetching and re-filtering
-		// bot-posted or already-replied comments on every poll cycle.
-		// When changes were detected but push failed, skip cursor advancement
-		// so the comments are retried on the next poll cycle.
-		if !changeDetected || pushed {
-			if task.maxCommentID > task.work.LastCommentID {
-				task.work.LastCommentID = task.maxCommentID
-			}
-			if task.maxReviewID > task.work.LastReviewID {
-				task.work.LastReviewID = task.maxReviewID
-			}
-		}
-	})
-}
-
-const maxCIFixAttempts = 3
-
-// ProcessCIFailures checks CI status for open PRs and invokes Claude to fix failures.
-func (a *Agent) ProcessCIFailures(ctx context.Context) {
-	a.emit(Event{
-		Type:   EventActionStarted,
-		Worker: a.workerName(),
-		State:  "working",
-		Action: "Checking CI status",
-	})
-	defer a.emit(Event{
-		Type:   EventActionCompleted,
-		Worker: a.workerName(),
-		State:  "idle",
-		Action: "CI check complete",
-	})
-	// Sequential phase: GitHub API calls, check run fetching, worktree setup
-	var tasks []ciTask
-
-	for _, work := range a.state.ActiveIssues {
-		if work.PRNumber == 0 || work.Status != StatusPROpen {
-			continue
-		}
-
-		headSHA, err := a.gh.GetPRHeadSHA(ctx, a.cfg.Owner, a.cfg.Repo, work.PRNumber)
-		if err != nil {
-			a.logger.Error("failed to get PR head SHA", "pr", work.PRNumber, "error", err)
-			continue
-		}
-
-		// Reset fix attempts counter if HEAD SHA changed (new commits pushed)
-		if work.LastCheckedCISHA != "" && work.LastCheckedCISHA != headSHA {
-			if work.CIFixAttempts > 0 {
-				a.logger.Info("new commits detected, resetting CI fix attempts", "pr", work.PRNumber, "old_sha", shortSHA(work.LastCheckedCISHA), "new_sha", shortSHA(headSHA), "previous_attempts", work.CIFixAttempts)
-				work.CIFixAttempts = 0
-				work.LastCIStatus = ""
-			}
-		}
-
-		if work.CIFixAttempts >= maxCIFixAttempts {
-			if work.LastCIStatus != "max-retries-reached" {
-				a.logger.Warn("CI fix attempts exhausted", "pr", work.PRNumber, "attempts", work.CIFixAttempts)
-				work.LastCIStatus = "max-retries-reached"
-			}
-			continue
-		}
-
-		runs, err := a.gh.GetCheckRuns(ctx, a.cfg.Owner, a.cfg.Repo, headSHA)
-		if err != nil {
-			a.logger.Error("failed to get check runs", "pr", work.PRNumber, "error", err)
-			continue
-		}
-
-		// Also query commit statuses (e.g. Prow) and merge into results
-		statusFailures, err := a.gh.GetCommitStatuses(ctx, a.cfg.Owner, a.cfg.Repo, headSHA)
-		if err != nil {
-			a.logger.Warn("failed to get commit statuses", "pr", work.PRNumber, "error", err)
-			// Non-fatal: continue with check runs only
-		} else {
-			runs = append(runs, statusFailures...)
-		}
-
-		// Collect completed failures — act immediately without waiting for all checks
-		var failures []CheckRun
-		allCompleted := true
-		for _, r := range runs {
-			if r.Status == "completed" && r.Conclusion == "failure" {
-				failures = append(failures, r)
-			}
-			if r.Status != "completed" {
-				allCompleted = false
-			}
-		}
-
-		if len(runs) == 0 {
-			// No check runs registered yet — don't mark as checked.
-			// This avoids the vacuous-truth bug where allCompleted is true
-			// (its initial value) because the loop over runs never executed.
-			a.logger.Debug("no check runs registered yet, waiting", "pr", work.PRNumber, "sha", shortSHA(headSHA))
-			continue
-		}
-		if len(failures) == 0 {
-			// If all checks completed with no failures and this SHA was already
-			// checked, skip. Only set LastCheckedCISHA when all checks are done
-			// to avoid skipping late-finishing failures.
-			if allCompleted && work.LastCheckedCISHA != headSHA {
-				work.LastCheckedCISHA = headSHA
-			}
-			continue
-		}
-
-		// Fast path: skip if this SHA was fully checked (all completed, no fix attempts)
-		// and there are no uninvestigated failures
-		if work.LastCheckedCISHA == headSHA && work.CIFixAttempts == 0 && allCompleted {
-			continue
-		}
-
-		// Filter out failures already investigated for this SHA.
-		// This prevents re-investigating a check that was already classified
-		// as RELATED or UNRELATED in a previous poll cycle.
-		var newFailures []CheckRun
-		for _, f := range failures {
-			if a.alreadyCheckedCI(ctx, work.PRNumber, headSHA, f.Name) {
-				a.logger.Info("CI already investigated for this check, skipping", "pr", work.PRNumber, "sha", shortSHA(headSHA), "check", f.Name)
-				continue
-			}
-			newFailures = append(newFailures, f)
-		}
-
-		if len(newFailures) == 0 {
-			work.LastCheckedCISHA = headSHA
-			continue
-		}
-
-		// Fetch logs for each failing check when output is missing or too short.
-		// Threshold of 50 chars filters out generic GitHub Actions boilerplate
-		// (e.g., "Process completed with exit code 1") that doesn't provide
-		// enough context for meaningful analysis.
-		// Skip entries with ID==0: these are commit-status entries (e.g. Prow)
-		// where Output contains a target_url, not log text, and no check-run
-		// log is available via the GitHub Actions API.
-		for i, f := range newFailures {
-			if f.ID == 0 {
-				continue
-			}
-			trimmed := strings.TrimSpace(f.Output)
-			if len(trimmed) < 50 {
-				log, err := a.gh.GetCheckRunLog(ctx, a.cfg.Owner, a.cfg.Repo, f.ID)
-				if err != nil {
-					a.logger.Warn("failed to get check run log", "check", f.Name, "error", err)
-				} else if log != "" {
-					newFailures[i].Output = log
-				}
-			}
-		}
-
-		a.logger.Info("CI failing, investigating", "pr", work.PRNumber, "failures", len(newFailures), "attempt", work.CIFixAttempts+1)
-
-		if err := a.ensureWorktreeReady(ctx, work); err != nil {
-			a.logger.Error("failed to prepare worktree", "pr", work.PRNumber, "error", err)
-			continue
-		}
-
-		// Get PR diff to help Claude determine if failure is related
-		diffOut, _, _ := a.runner.Run(ctx, work.WorktreePath, "git", "diff", "--stat", a.originDefaultBranch())
-		diff := string(diffOut)
-
-		// Get commits in the PR to help Claude identify which commit introduced the failure
-		commits := a.getPRCommits(ctx, work.WorktreePath)
-
-		// Create one task per failure so each gets its own RELATED/UNRELATED classification.
-		// This prevents a single unrelated failure (e.g. a policy check) from masking a
-		// related failure (e.g. a unit test) when they fail on the same SHA.
-		for _, f := range newFailures {
-			tasks = append(tasks, ciTask{
-				work:     work,
-				headSHA:  headSHA,
-				failures: []CheckRun{f},
-				diff:     diff,
-				commits:  commits,
-			})
-		}
-	}
-
-	// Parallel phase: Claude invocations and post-processing
-	runSequential(ctx, tasks, func(ctx context.Context, task ciTask) {
-		prompt := buildCIFixPrompt(*task.work, task.failures, task.diff, task.commits, a.cfg.SignedOffBy, a.cfg.SkipFix)
-		result, err := a.codeAgent.Run(ctx, a.runner, task.work.WorktreePath, prompt, a.logger, true)
-		if err != nil {
-			a.logger.Error("agent failed to investigate CI", "pr", task.work.PRNumber, "error", err)
-			task.work.CIFixAttempts++
-			task.work.LastCIStatus = "failure"
-			task.work.LastCheckedCISHA = task.headSHA
-			return
-		}
-
-		// Strip markdown formatting (bold, italic) before checking prefix
-		cleaned := strings.TrimLeft(strings.TrimSpace(result.Result), "*_")
-
-		// Check if the response contains UNRELATED, INFRASTRUCTURE, or RELATED.
-		// The agent often puts detailed analysis before the keyword, sometimes
-		// many lines deep. Scan the entire response for the keyword on any line.
-		// Priority: INFRASTRUCTURE > UNRELATED > RELATED (check in this order
-		// so RELATED doesn't match the "UNRELATED" substring).
-		foundKeyword := ""
-		for line := range strings.SplitSeq(cleaned, "\n") {
-			trimmed := strings.TrimLeft(strings.TrimSpace(line), "*_-—:>")
-			switch {
-			case strings.HasPrefix(trimmed, "INFRASTRUCTURE"):
-				foundKeyword = "INFRASTRUCTURE"
-			case strings.HasPrefix(trimmed, "UNRELATED") && foundKeyword == "":
-				foundKeyword = "UNRELATED"
-			case strings.HasPrefix(trimmed, "RELATED") && foundKeyword == "":
-				foundKeyword = "RELATED"
-			}
-			if foundKeyword == "INFRASTRUCTURE" {
-				break // highest priority, stop scanning
-			}
-		}
-
-		if foundKeyword == "" {
-			a.logger.Warn("agent response did not contain UNRELATED, INFRASTRUCTURE, or RELATED, skipping to avoid noise",
-				"pr", task.work.PRNumber,
-				"response_preview", truncateString(cleaned, 200))
-			task.work.CIFixAttempts++
-			task.work.LastCIStatus = "investigation-inconclusive"
-			task.work.LastCheckedCISHA = task.headSHA
-			return
-		}
-
-		if foundKeyword == "INFRASTRUCTURE" {
-			a.logger.Info("CI failure is an infrastructure issue", "pr", task.work.PRNumber)
-			// Extract explanation: everything after INFRASTRUCTURE in the full response,
-			// stripping any separator characters (e.g. "INFRASTRUCTURE: ..." or "INFRASTRUCTURE — ...").
-			idx := strings.Index(cleaned, "INFRASTRUCTURE")
-			explanation := strings.TrimSpace(strings.TrimLeft(strings.TrimPrefix(cleaned[idx:], "INFRASTRUCTURE"), " :—–-"))
-			marker := ciMarker(task.headSHA, task.failures[0].Name)
-			if a.ShouldSkipComment("ci-infrastructure") {
-				a.logger.Info("skipping CI infrastructure comment (--skip-comment)", "pr", task.work.PRNumber)
-				a.markCIChecked(task.work, task.headSHA, task.failures[0].Name)
-			} else {
-				comment := fmt.Sprintf("CI check `%s` failed on commit %s due to an infrastructure issue (not a flaky test).", task.failures[0].Name, shortSHA(task.headSHA))
-				if explanation != "" {
-					comment += "\n\n" + explanation
-				}
-				comment += "\n\n" + marker
-				if err := a.gh.AddIssueComment(ctx, a.cfg.Owner, a.cfg.Repo, task.work.PRNumber, comment); err != nil {
-					a.logger.Error("failed to post CI infrastructure comment", "pr", task.work.PRNumber, "error", err)
-					task.work.LastCIStatus = "investigation-inconclusive"
-					return
-				}
-			}
-			task.work.LastCIStatus = "infrastructure-failure"
-			task.work.LastCheckedCISHA = task.headSHA
-			// Do NOT create a flaky issue — infrastructure failures are transient
-			return
-		}
-
-		if foundKeyword == "UNRELATED" {
-			a.logger.Info("CI failure is unrelated to PR changes", "pr", task.work.PRNumber)
-			// Extract explanation: everything after UNRELATED in the full response,
-			// stripping any separator characters (e.g. "UNRELATED: ..." or "UNRELATED — ...").
-			idx := strings.Index(cleaned, "UNRELATED")
-			explanation := strings.TrimSpace(strings.TrimLeft(strings.TrimPrefix(cleaned[idx:], "UNRELATED"), " :—–-"))
-			marker := ciMarker(task.headSHA, task.failures[0].Name)
-			if a.ShouldSkipComment("ci-unrelated") {
-				a.logger.Info("skipping CI unrelated comment (--skip-comment)", "pr", task.work.PRNumber)
-				a.markCIChecked(task.work, task.headSHA, task.failures[0].Name)
-			} else {
-				comment := fmt.Sprintf("CI check `%s` failed on commit %s but appears unrelated to this PR's changes.", task.failures[0].Name, shortSHA(task.headSHA))
-				if explanation != "" {
-					comment += "\n\n" + explanation
-				}
-				comment += "\n\n" + marker
-				if err := a.gh.AddIssueComment(ctx, a.cfg.Owner, a.cfg.Repo, task.work.PRNumber, comment); err != nil {
-					a.logger.Error("failed to post CI unrelated comment", "pr", task.work.PRNumber, "error", err)
-					task.work.LastCIStatus = "investigation-inconclusive"
-					return
-				}
-			}
-			task.work.LastCIStatus = "unrelated-failure"
-			task.work.LastCheckedCISHA = task.headSHA
-
-			// Create a flaky CI issue if configured
-			if a.cfg.CreateFlakyIssues {
-				// Skip flaky issue creation if both the check run output and the
-				// agent's explanation are too short. The agent may have gathered
-				// useful context from its own investigation even when the check
-				// run output was empty.
-				trimmedOutput := strings.TrimSpace(task.failures[0].Output)
-				if len(trimmedOutput) < 50 && len(explanation) < 50 {
-					a.logger.Warn("skipping flaky issue creation: insufficient context",
-						"pr", task.work.PRNumber,
-						"check", task.failures[0].Name,
-						"output_length", len(trimmedOutput),
-						"explanation_length", len(explanation))
-					return
-				}
-
-				failingTest := parseFailingTest(explanation)
-				var issueTitle string
-				if failingTest != "" {
-					issueTitle = fmt.Sprintf("Flaky CI: %s / %s", task.failures[0].Name, failingTest)
-				} else {
-					issueTitle = fmt.Sprintf("Flaky CI: %s", task.failures[0].Name)
-				}
-
-				// Search for existing open issues with the flaky label
-				searchQuery := fmt.Sprintf("repo:%s/%s is:issue is:open label:%s",
-					a.cfg.Owner, a.cfg.Repo, a.cfg.FlakyLabel)
-				existingIssues, err := a.gh.SearchIssues(ctx, searchQuery)
-
-				var issueNum int
-				if err != nil {
-					a.logger.Warn("failed to search for existing flaky issues", "error", err)
-				} else if len(existingIssues) > 0 {
-					// Fast path: exact title match skips LLM matching entirely.
-					// This handles the most common duplicate case (same check name)
-					// and prevents duplicates when the same check fails on multiple
-					// PRs in the same poll cycle.
-					for _, existing := range existingIssues {
-						if existing.Title == issueTitle {
-							issueNum = existing.Number
-							a.logger.Info("exact title match for existing flaky CI issue", "issue", issueNum, "check", task.failures[0].Name)
-							break
-						}
-					}
-
-					// If no exact title match, ask the agent for root-cause matching.
-					if issueNum == 0 {
-						// Use the check run output for matching, falling back to the
-						// agent's explanation when the output is empty/short.
-						matchOutput := task.failures[0].Output
-						if len(strings.TrimSpace(matchOutput)) < 50 {
-							matchOutput = explanation
-						}
-						matchPrompt := buildFlakyMatchPrompt(task.failures[0].Name, matchOutput, existingIssues)
-						matchResult, matchErr := a.codeAgent.Run(ctx, a.runner, task.work.WorktreePath, matchPrompt, a.logger, false)
-						if matchErr != nil {
-							a.logger.Warn("failed to run agent for flaky issue matching", "error", matchErr)
-						} else {
-							matchResponse := strings.TrimSpace(matchResult.Result)
-							if matchedNum, ok := parseFlakyMatch(matchResponse); ok {
-								for _, existing := range existingIssues {
-									if existing.Number == matchedNum {
-										issueNum = matchedNum
-										break
-									}
-								}
-								if issueNum > 0 {
-									a.logger.Info("agent matched existing flaky CI issue", "issue", issueNum, "check", task.failures[0].Name)
-								} else {
-									a.logger.Warn("agent returned MATCH for unknown issue", "matched_issue", matchedNum, "check", task.failures[0].Name)
-								}
-							}
-						}
-					}
-
-					if issueNum > 0 {
-						if !a.ShouldSkipComment("flaky") {
-							if err := a.gh.AddIssueComment(ctx, a.cfg.Owner, a.cfg.Repo, task.work.PRNumber,
-								fmt.Sprintf("This appears to be a duplicate of existing flaky test issue #%d.\n\n%s", issueNum, a.botComment())); err != nil {
-								a.logger.Error("failed to post existing flaky issue reference comment", "pr", task.work.PRNumber, "error", err)
-							}
-						}
-						return
-					}
-				}
-
-				// No existing issue matched, create a new one
-				// Use the failing test name in the body if available, otherwise fall back to lane name
-				testNameForBody := task.failures[0].Name
-				if failingTest != "" {
-					testNameForBody = failingTest
-				}
-				// Strip the FAILING_TEST: line from the explanation so it doesn't appear as raw text in the issue body
-				cleanedExplanation := explanation
-				if failingTest != "" {
-					var lines []string
-					for line := range strings.SplitSeq(explanation, "\n") {
-						if !strings.HasPrefix(strings.TrimSpace(line), "FAILING_TEST:") {
-							lines = append(lines, line)
-						}
-					}
-					cleanedExplanation = strings.TrimSpace(strings.Join(lines, "\n"))
-				}
-				issueBody := fmt.Sprintf("### Which jobs are flaking?\n\n"+
-					"%s\n\n"+
-					"Detected in PR #%d, commit %s.\n\n"+
-					"### Which tests are flaking?\n\n"+
-					"%s\n\n"+
-					"### Since when has it been flaking?\n\n"+
-					"%s\n\n"+
-					"### Reason for failure (if possible)\n\n"+
-					"%s\n\n"+
-					"### Anything else we need to know?\n\n"+
-					"Automatically created by [oompa](https://github.com/qinqon/oompa).\n\n"+
-					"%s",
-					task.failures[0].Name, task.work.PRNumber, shortSHA(task.headSHA),
-					testNameForBody, time.Now().Format("2006-01-02"), cleanedExplanation,
-					a.botComment())
-				issueNum, err = a.gh.CreateIssue(ctx, a.cfg.Owner, a.cfg.Repo, issueTitle, issueBody, []string{a.cfg.FlakyLabel})
-				if err != nil {
-					a.logger.Error("failed to create flaky CI issue", "error", err)
-				} else {
-					a.logger.Info("created flaky CI issue", "issue", issueNum)
-					if !a.ShouldSkipComment("flaky") {
-						if err := a.gh.AddIssueComment(ctx, a.cfg.Owner, a.cfg.Repo, task.work.PRNumber,
-							fmt.Sprintf("Opened issue #%d to track this flaky test.\n\n%s", issueNum, a.botComment())); err != nil {
-							a.logger.Error("failed to post flaky issue reference comment", "pr", task.work.PRNumber, "error", err)
-						}
-					}
-				}
-			}
-
-			return
-		}
-
-		// Claude said RELATED
-		if a.cfg.SkipFix {
-			// Skip-fix mode: just post the analysis, don't try to fix or push
-			a.logger.Info("CI failure is related (skip-fix mode, not pushing)", "pr", task.work.PRNumber)
-			idx := strings.Index(cleaned, "RELATED")
-			analysis := strings.TrimPrefix(cleaned[idx:], "RELATED")
-			analysis = strings.TrimSpace(analysis)
-			marker := ciMarker(task.headSHA, task.failures[0].Name)
-			if a.ShouldSkipComment("ci-related") {
-				a.logger.Info("skipping CI related comment (--skip-comment)", "pr", task.work.PRNumber)
-				a.markCIChecked(task.work, task.headSHA, task.failures[0].Name)
-			} else {
-				if err := a.gh.AddIssueComment(ctx, a.cfg.Owner, a.cfg.Repo, task.work.PRNumber,
-					fmt.Sprintf("CI check `%s` is failing on commit %s and appears related to this PR's changes.\n\n%s\n\n%s", task.failures[0].Name, shortSHA(task.headSHA), analysis, marker)); err != nil {
-					a.logger.Error("failed to post CI analysis comment", "pr", task.work.PRNumber, "error", err)
-				}
-			}
-			task.work.LastCIStatus = "related-skip-fix"
-			task.work.LastCheckedCISHA = task.headSHA
-			return
-		}
-
-		// Check if there are fixup commits, amended commits, or uncommitted changes
-		pushed := false
-		hasFixupCommits := a.hasFixupCommits(ctx, task.work.WorktreePath)
-		hasUncommitted := a.hasUncommittedChanges(ctx, task.work.WorktreePath)
-
-		switch {
-		case hasFixupCommits:
-			// Run autosquash rebase to merge fixup commits into their targets
-			if err := a.gitAutosquashRebase(ctx, task.work.WorktreePath); err != nil {
-				a.logger.Error("failed to autosquash fixup commits", "pr", task.work.PRNumber, "error", err)
-			} else if err := a.gitPush(ctx, task.work.WorktreePath, true); err != nil {
-				a.logger.Error("failed to push CI fix", "pr", task.work.PRNumber, "error", err)
-			} else {
-				pushed = true
-			}
-		case hasUncommitted:
-			// Claude left uncommitted changes — amend them into HEAD as fallback
-			if err := a.gitAmendAll(ctx, task.work.WorktreePath); err != nil {
-				a.logger.Error("failed to amend commit for CI fix", "pr", task.work.PRNumber, "error", err)
-			} else if err := a.gitPush(ctx, task.work.WorktreePath, true); err != nil {
-				a.logger.Error("failed to push CI fix", "pr", task.work.PRNumber, "error", err)
-			} else {
-				pushed = true
-			}
-		default:
-			// No fixups and no uncommitted changes — Claude may have amended
-			// the commit directly. Check if HEAD differs from the remote SHA.
-			localSHA, _, err := a.runner.Run(ctx, task.work.WorktreePath, "git", "rev-parse", "HEAD")
-			if err == nil && strings.TrimSpace(string(localSHA)) != task.headSHA {
-				a.logger.Info("Claude amended commit directly, pushing", "pr", task.work.PRNumber)
-				if err := a.gitPush(ctx, task.work.WorktreePath, true); err != nil {
-					a.logger.Error("failed to push CI fix", "pr", task.work.PRNumber, "error", err)
-				} else {
-					pushed = true
-				}
-			}
-		}
-
-		// After pushing (or not pushing), fetch the current HEAD SHA to update state
-		currentHeadSHA := task.headSHA
-		if pushed {
-			// Get the new HEAD SHA after the push
-			newHeadSHA, err := a.gh.GetPRHeadSHA(ctx, a.cfg.Owner, a.cfg.Repo, task.work.PRNumber)
-			if err != nil {
-				a.logger.Warn("failed to get new HEAD SHA after push", "pr", task.work.PRNumber, "error", err)
-				// Fall back to old SHA if we can't get the new one
-			} else {
-				currentHeadSHA = newHeadSHA
-			}
-		}
-
-		marker := ciMarker(task.headSHA, task.failures[0].Name)
-		if pushed {
-			a.logger.Info("CI failure is related, pushed a fix", "pr", task.work.PRNumber)
-			if a.ShouldSkipComment("ci-related") {
-				a.logger.Info("skipping CI fix-pushed comment (--skip-comment)", "pr", task.work.PRNumber)
-				a.markCIChecked(task.work, task.headSHA, task.failures[0].Name)
-			} else {
-				if err := a.gh.AddIssueComment(ctx, a.cfg.Owner, a.cfg.Repo, task.work.PRNumber,
-					fmt.Sprintf("CI check `%s` was failing on commit %s. Pushed a fix.\n\n%s", task.failures[0].Name, shortSHA(task.headSHA), marker)); err != nil {
-					a.logger.Error("failed to post CI fix comment", "pr", task.work.PRNumber, "error", err)
-				}
-			}
-		} else {
-			a.logger.Warn("Claude said RELATED but no changes to push", "pr", task.work.PRNumber)
-			if a.ShouldSkipComment("ci-related") {
-				a.logger.Info("skipping CI fix-failed comment (--skip-comment)", "pr", task.work.PRNumber)
-				a.markCIChecked(task.work, task.headSHA, task.failures[0].Name)
-			} else {
-				if err := a.gh.AddIssueComment(ctx, a.cfg.Owner, a.cfg.Repo, task.work.PRNumber,
-					fmt.Sprintf("CI check `%s` is failing on commit %s. Investigated but could not push a fix.\n\n%s", task.failures[0].Name, shortSHA(task.headSHA), marker)); err != nil {
-					a.logger.Error("failed to post CI investigation comment", "pr", task.work.PRNumber, "error", err)
-				}
-			}
-		}
-		task.work.CIFixAttempts++
-		task.work.LastCIStatus = "failure"
-		task.work.LastCheckedCISHA = currentHeadSHA
-	})
-}
-
-// ProcessConflicts checks for merge conflicts (dirty mergeable_state) and tries to resolve them.
-// For simple rebases when a PR is just behind the base branch, use ProcessRebase instead.
-func (a *Agent) ProcessConflicts(ctx context.Context) {
-	a.emit(Event{
-		Type:   EventActionStarted,
-		Worker: a.workerName(),
-		State:  "working",
-		Action: "Checking for merge conflicts",
-	})
-	defer a.emit(Event{
-		Type:   EventActionCompleted,
-		Worker: a.workerName(),
-		State:  "idle",
-		Action: "Conflict check complete",
-	})
-	// Sequential phase: GitHub API calls, worktree setup, git fetch, automatic rebase attempts
-	var tasks []conflictTask
-
-	for _, work := range a.state.ActiveIssues {
-		if work.PRNumber == 0 || work.Status != StatusPROpen {
-			continue
-		}
-
-		mergeState, err := a.gh.GetPRMergeable(ctx, a.cfg.Owner, a.cfg.Repo, work.PRNumber)
-		if err != nil {
-			a.logger.Error("failed to get PR mergeable state", "pr", work.PRNumber, "error", err)
-			continue
-		}
-
-		a.logger.Debug("PR mergeable state", "pr", work.PRNumber, "state", mergeState)
-
-		// Only handle actual merge conflicts
-		if mergeState != "dirty" {
-			continue
-		}
-
-		// Check if we already posted a conflict comment for the current HEAD
-		headSHA, _ := a.gh.GetPRHeadSHA(ctx, a.cfg.Owner, a.cfg.Repo, work.PRNumber)
-		if headSHA != "" && a.hasExistingBotComment(ctx, work.PRNumber, "rebase") && a.hasExistingBotComment(ctx, work.PRNumber, shortSHA(headSHA)) {
-			continue
-		}
-
-		a.logger.Info("PR needs rebase, attempting", "pr", work.PRNumber, "mergeable_state", mergeState)
-
-		if err := a.ensureWorktreeReady(ctx, work); err != nil {
-			a.logger.Error("failed to prepare worktree", "pr", work.PRNumber, "error", err)
-			continue
-		}
-
-		// Fetch all remotes and try automatic rebase against the upstream default branch
-		a.runner.Run(ctx, work.WorktreePath, "git", "fetch", "--all") //nolint:errcheck // best-effort
-
-		// Try automatic rebase (with retry on unstaged changes)
-		stderr, rebaseErr := a.gitRebaseWithRetry(ctx, work.WorktreePath, work.PRNumber)
-		if rebaseErr == nil {
-			// Rebase succeeded, force push
-			if pushErr := a.gitPush(ctx, work.WorktreePath, true); pushErr != nil {
-				a.logger.Error("failed to push after rebase", "pr", work.PRNumber, "error", pushErr)
-			} else {
-				a.logger.Info("rebased and pushed successfully", "pr", work.PRNumber)
-				if !a.ShouldSkipComment("conflict") {
-					_ = a.gh.AddIssueComment(ctx, a.cfg.Owner, a.cfg.Repo, work.PRNumber,
-						fmt.Sprintf("Rebased commit %s on main and pushed.\n\n%s", shortSHA(headSHA), a.botComment()))
-				}
-			}
-			continue
-		}
-
-		// Rebase failed — abort and let Claude try
-		a.runner.Run(ctx, work.WorktreePath, "git", "rebase", "--abort") //nolint:errcheck // best-effort
-		a.logger.Info("automatic rebase failed, invoking Claude to resolve conflicts", "pr", work.PRNumber, "stderr", stderr)
-
-		tasks = append(tasks, conflictTask{
-			work:         work,
-			headSHA:      headSHA,
-			rebaseErr:    rebaseErr,
-			rebaseStderr: stderr,
-		})
-	}
-
-	// Parallel phase: Claude invocations for conflict resolution
-	a.resolveConflictsParallel(ctx, tasks)
-}
-
-// ProcessRebase rebases PRs that are behind the base branch but have no merge conflicts.
-// For PRs with actual merge conflicts (dirty state), use ProcessConflicts instead.
-// If a rebase fails due to conflicts, this delegates to the conflict resolution flow.
-func (a *Agent) ProcessRebase(ctx context.Context) {
-	a.emit(Event{
-		Type:   EventActionStarted,
-		Worker: a.workerName(),
-		State:  "rebasing",
-		Action: "Checking for outdated PRs",
-	})
-	defer a.emit(Event{
-		Type:   EventActionCompleted,
-		Worker: a.workerName(),
-		State:  "idle",
-		Action: "Rebase check complete",
-	})
-	// Sequential phase: check states, try automatic rebase, collect failed conflicts into tasks
-	var tasks []conflictTask
-
-	for _, work := range a.state.ActiveIssues {
-		if work.PRNumber == 0 || work.Status != StatusPROpen {
-			continue
-		}
-
-		mergeState, err := a.gh.GetPRMergeable(ctx, a.cfg.Owner, a.cfg.Repo, work.PRNumber)
-		if err != nil {
-			a.logger.Error("failed to get PR mergeable state", "pr", work.PRNumber, "error", err)
-			continue
-		}
-
-		a.logger.Debug("PR mergeable state", "pr", work.PRNumber, "state", mergeState)
-
-		needsRebase := mergeState == "behind"
-
-		// When mergeable_state is something else (e.g. "unstable", "blocked"), it may mask
-		// the branch being behind. Use the compare API as a fallback.
-		if !needsRebase && mergeState != "dirty" {
-			behind, err := a.gh.IsPRBehind(ctx, a.cfg.Owner, a.cfg.Repo, work.PRNumber)
-			if err != nil {
-				a.logger.Warn("failed to check if PR is behind", "pr", work.PRNumber, "error", err)
-			}
-			needsRebase = behind
-		}
-
-		if !needsRebase {
-			continue
-		}
-
-		headSHA, _ := a.gh.GetPRHeadSHA(ctx, a.cfg.Owner, a.cfg.Repo, work.PRNumber)
-		if headSHA != "" && a.hasExistingBotComment(ctx, work.PRNumber, "rebase") && a.hasExistingBotComment(ctx, work.PRNumber, shortSHA(headSHA)) {
-			continue
-		}
-
-		a.logger.Info("PR is behind base branch, rebasing", "pr", work.PRNumber, "mergeable_state", mergeState)
-
-		if err := a.ensureWorktreeReady(ctx, work); err != nil {
-			a.logger.Error("failed to prepare worktree", "pr", work.PRNumber, "error", err)
-			continue
-		}
-
-		a.runner.Run(ctx, work.WorktreePath, "git", "fetch", "--all") //nolint:errcheck // best-effort
-
-		stderr, rebaseErr := a.gitRebaseWithRetry(ctx, work.WorktreePath, work.PRNumber)
-		if rebaseErr != nil {
-			a.runner.Run(ctx, work.WorktreePath, "git", "rebase", "--abort") //nolint:errcheck // best-effort
-
-			// Check if the rebase failed due to conflicts
-			if isConflictError(stderr) {
-				a.logger.Info("rebase failed with conflicts, will invoke conflict resolution", "pr", work.PRNumber)
-				tasks = append(tasks, conflictTask{
-					work:         work,
-					headSHA:      headSHA,
-					rebaseErr:    rebaseErr,
-					rebaseStderr: stderr,
-				})
-			} else {
-				// Non-conflict rebase failure (e.g., corrupt repo state, hook failure)
-				a.logger.Error("rebase failed for non-conflict reason", "pr", work.PRNumber, "stderr", stderr)
-			}
-			continue
-		}
-
-		if pushErr := a.gitPush(ctx, work.WorktreePath, true); pushErr != nil {
-			a.logger.Error("failed to push after rebase", "pr", work.PRNumber, "error", pushErr)
-		} else {
-			a.logger.Info("rebased and pushed successfully", "pr", work.PRNumber)
-			if !a.ShouldSkipComment("rebase") {
-				_ = a.gh.AddIssueComment(ctx, a.cfg.Owner, a.cfg.Repo, work.PRNumber,
-					fmt.Sprintf("Rebased commit %s on main and pushed.\n\n%s", shortSHA(headSHA), a.botComment()))
-			}
-		}
-	}
-
-	// Parallel phase: invoke Claude for conflict resolution on collected tasks
-	a.resolveConflictsParallel(ctx, tasks)
-}
-
-// triageLookbackRunLimit is the maximum number of runs to fetch when scanning a lookback window.
-const triageLookbackRunLimit = 50
-
-// triageRunTask pairs a CI source with a run for parallel investigation.
-type triageRunTask struct {
-	ciSource CIJobSource
-	run      JobRun
-}
-
-// ProcessTriageJobs monitors periodic CI jobs for failures and investigates them.
-func (a *Agent) ProcessTriageJobs(ctx context.Context) {
-	if len(a.cfg.TriageJobs) == 0 {
-		return
-	}
-
-	var tasks []triageRunTask
-
-	for _, jobURL := range a.cfg.TriageJobs {
-		a.logger.Debug("processing triage job", "url", jobURL)
-
-		// Parse the CI job URL to determine the backend
-		ciSource, err := ParseCIJobURL(jobURL, a.gh)
-		if err != nil {
-			a.logger.Error("failed to parse CI job URL", "url", jobURL, "error", err)
-			continue
-		}
-
-		// Fetch more runs when scanning a time window
-		limit := 5
-		if a.cfg.TriageLookback > 0 {
-			limit = triageLookbackRunLimit
-		}
-
-		runs, err := ciSource.ListRecentRuns(ctx, limit)
-		if err != nil {
-			a.logger.Error("failed to list recent runs", "job", ciSource.JobName(), "error", err)
-			continue
-		}
-
-		if len(runs) == 0 {
-			a.logger.Debug("no recent runs found", "job", ciSource.JobName())
-			continue
-		}
-
-		// Determine which runs to process
-		var cutoff time.Time
-		if a.cfg.TriageLookback > 0 {
-			cutoff = time.Now().Add(-a.cfg.TriageLookback)
-		}
-
-		runsToProcess := runs
-		if cutoff.IsZero() {
-			// Default: only the most recent run
-			runsToProcess = runs[:1]
-		}
-
-		for _, run := range runsToProcess {
-			// Stop once we hit runs older than the lookback window (runs are sorted descending)
-			if !cutoff.IsZero() && run.Timestamp.Before(cutoff) {
-				break
-			}
-
-			// Skip if already investigated
-			if a.state.IsRunInvestigated(ciSource.JobName(), run.ID) {
-				a.logger.Debug("run already investigated", "job", ciSource.JobName(), "runID", run.ID)
-				continue
-			}
-
-			// Skip if the run passed (mark as investigated)
-			if run.Status == "success" {
-				a.logger.Info("run passed, skipping", "job", ciSource.JobName(), "runID", run.ID)
-				a.state.MarkRunInvestigated(ciSource.JobName(), run.ID)
-				continue
-			}
-
-			a.logger.Info("investigating failed run", "job", ciSource.JobName(), "runID", run.ID, "status", run.Status)
-			tasks = append(tasks, triageRunTask{ciSource: ciSource, run: run})
-		}
-	}
-
-	// Investigate collected failed runs in parallel
-	runSequential(ctx, tasks, func(ctx context.Context, task triageRunTask) {
-		a.investigateTriageRun(ctx, task.ciSource, task.run)
-	})
-}
-
-// investigateTriageRun handles the investigation of a single failed CI run.
-func (a *Agent) investigateTriageRun(ctx context.Context, ciSource CIJobSource, run JobRun) {
-	// Fetch build log
-	buildLog, err := ciSource.FetchLog(ctx, run.ID)
-	if err != nil {
-		a.logger.Error("failed to fetch build log", "job", ciSource.JobName(), "runID", run.ID, "error", err)
-		return
-	}
-
-	// Truncate log if too large (keep last 50KB to focus on recent failures)
-	const maxLogSize = 50000
-	if len(buildLog) > maxLogSize {
-		buildLog = "...\n[Log truncated, showing last 50KB]\n...\n\n" + buildLog[len(buildLog)-maxLogSize:]
-	}
-
-	// Create a worktree on the default branch for read-only codebase access
-	branchName := fmt.Sprintf("triage/%s", run.ID)
-
-	// Ensure repo is cloned
-	if err := a.worktrees.EnsureRepoCloned(ctx); err != nil {
-		a.logger.Error("failed to ensure repo cloned", "error", err)
-		return
-	}
-
-	// Create worktree on default branch
-	worktreePath, err := a.worktrees.CreateWorktree(ctx, branchName)
-	if err != nil {
-		a.logger.Error("failed to create triage worktree", "error", err)
-		return
-	}
-
-	// Checkout the default branch (CreateWorktree creates a new branch, we want default)
-	a.runner.Run(ctx, worktreePath, "git", "checkout", a.defaultBranch()) //nolint:errcheck // best-effort
-
-	// Build the triage prompt
-	prompt := buildPeriodicCITriagePrompt(ciSource.JobName(), run.ID, buildLog, a.cfg.Owner, a.cfg.Repo)
-
-	// Run agent in the worktree
-	a.logger.Info("running agent for CI triage", "job", ciSource.JobName(), "runID", run.ID)
-	result, err := a.codeAgent.Run(ctx, a.runner, worktreePath, prompt, a.logger, false)
-	if err != nil {
-		a.logger.Error("agent failed during CI triage", "job", ciSource.JobName(), "runID", run.ID, "error", err)
-		_ = a.worktrees.RemoveWorktree(ctx, worktreePath)
-		return
-	}
-
-	analysis := result.Result
-	a.logger.Info("CI triage analysis complete", "job", ciSource.JobName(), "runID", run.ID)
-	a.logger.Debug("analysis output", "output", analysis)
-
-	// If --create-flaky-issues is set, create a GitHub issue with the analysis
-	if a.cfg.CreateFlakyIssues {
-		a.logger.Info("creating issue for CI failure", "job", ciSource.JobName(), "runID", run.ID)
-
-		// Search for existing issues about this job to avoid duplicates
-		searchQuery := fmt.Sprintf("repo:%s/%s is:issue is:open in:title %q", a.cfg.Owner, a.cfg.Repo, ciSource.JobName())
-		existingIssues, err := a.gh.SearchIssues(ctx, searchQuery)
-		if err != nil {
-			a.logger.Warn("failed to search for existing issues", "error", err)
-		}
-
-		if len(existingIssues) > 0 {
-			a.logger.Info("found existing issue for this job, skipping issue creation", "job", ciSource.JobName(), "issue", existingIssues[0].Number)
-		} else {
-			// Create a new issue
-			title := fmt.Sprintf("CI Failure: %s", ciSource.JobName())
-			body := fmt.Sprintf(`Periodic CI job **%s** failed in run [%s](%s).
-
-## Analysis
-
-%s
-
----
-*This issue was automatically created by oompa based on CI failure analysis.*
-<!-- oompa-triage -->`, ciSource.JobName(), run.ID, run.LogURL, analysis)
-
-			issueNumber, err := a.gh.CreateIssue(ctx, a.cfg.Owner, a.cfg.Repo, title, body, []string{a.cfg.FlakyLabel})
-			if err != nil {
-				a.logger.Error("failed to create issue", "error", err)
-			} else {
-				a.logger.Info("created issue for CI failure", "job", ciSource.JobName(), "issue", issueNumber)
-			}
-		}
-	}
-
-	// Mark the run as investigated
-	a.state.MarkRunInvestigated(ciSource.JobName(), run.ID)
-
-	// Clean up the triage worktree
-	if err := a.worktrees.RemoveWorktree(ctx, worktreePath); err != nil {
-		a.logger.Warn("failed to remove triage worktree", "path", worktreePath, "error", err)
-	}
 }
 
 // CleanupDone removes worktrees for merged or closed PRs.
@@ -1714,238 +421,6 @@ func (a *Agent) pushRemote() string {
 	return "origin"
 }
 
-// buildPRBody constructs a PR description. If Claude wrote a .pr-body.md file
-// (filled from the repo's PR template), that is used. Otherwise falls back to
-// constructing a body from the git log.
-func (a *Agent) buildPRBody(ctx context.Context, worktreePath string, issueNumber int) string {
-	prBodyFile := filepath.Join(worktreePath, ".pr-body.md")
-	if content, err := os.ReadFile(prBodyFile); err == nil {
-		os.Remove(prBodyFile)
-		body := strings.TrimSpace(string(content))
-		if !strings.Contains(body, botMarker) {
-			body += "\n\n" + a.botComment()
-		}
-		return body
-	}
-
-	// Fallback: construct from git log body only (skip subject to avoid duplication).
-	logOut, _, _ := a.runner.Run(ctx, worktreePath, "git", "log", a.originDefaultBranch()+"..HEAD", "--format=%b---")
-	rawBody := strings.TrimSpace(string(logOut))
-
-	body := fmt.Sprintf("Fixes #%d\n\n", issueNumber)
-	if rawBody != "" {
-		body += stripSignedOffBy(rawBody) + "\n\n"
-	}
-	body += a.botComment()
-	return body
-}
-
-// stripSignedOffBy removes "Signed-off-by: ..." trailer lines from a string.
-func stripSignedOffBy(s string) string {
-	var kept []string
-	for line := range strings.SplitSeq(s, "\n") {
-		if strings.HasPrefix(strings.TrimSpace(line), "Signed-off-by:") {
-			continue
-		}
-		kept = append(kept, line)
-	}
-	return strings.TrimSpace(strings.Join(kept, "\n"))
-}
-
-// ensureWorktreeReady ensures the repo is cloned and the worktree exists for the given work item.
-// If the worktree was lost (e.g. after a restart with a fresh volume), it recreates it.
-func (a *Agent) ensureWorktreeReady(ctx context.Context, work *IssueWork) error {
-	if err := a.worktrees.EnsureRepoCloned(ctx); err != nil {
-		return err
-	}
-	worktreePath, err := a.worktrees.CreateWorktree(ctx, work.BranchName)
-	if err != nil {
-		return err
-	}
-	work.WorktreePath = worktreePath
-
-	// Pull the latest from the push remote
-	return a.worktrees.SyncWorktree(ctx, worktreePath)
-}
-
-// hasUncommittedChanges returns true if the worktree has staged or unstaged changes.
-func (a *Agent) hasUncommittedChanges(ctx context.Context, worktreePath string) bool {
-	out, _, _ := a.runner.Run(ctx, worktreePath, "git", "status", "--porcelain")
-	return strings.TrimSpace(string(out)) != ""
-}
-
-// gitAmendAll stages all changes and amends the current commit.
-func (a *Agent) gitAmendAll(ctx context.Context, worktreePath string) error {
-	if a.cfg.DryRun {
-		a.logger.Info("[dry-run] would amend commit", "worktree", worktreePath)
-		return nil
-	}
-	if _, stderr, err := a.runner.Run(ctx, worktreePath, "git", "add", "-A"); err != nil {
-		return fmt.Errorf("git add: %w (stderr: %s)", err, string(stderr))
-	}
-	if _, stderr, err := a.runner.Run(ctx, worktreePath, "git", "commit", "--amend", "--no-edit"); err != nil {
-		return fmt.Errorf("git commit --amend: %w (stderr: %s)", err, string(stderr))
-	}
-	return nil
-}
-
-// gitSquashInto squashes all commits after the given SHA into that commit.
-// Used to fold agent-created review feedback commits back into the original HEAD.
-func (a *Agent) gitSquashInto(ctx context.Context, worktreePath, targetSHA string) error {
-	if a.cfg.DryRun {
-		a.logger.Info("[dry-run] would squash into", "worktree", worktreePath, "target", shortSHA(targetSHA))
-		return nil
-	}
-	// Stage all changes first to capture any unstaged modifications the agent left behind
-	if _, stderr, err := a.runner.Run(ctx, worktreePath, "git", "add", "-A"); err != nil {
-		return fmt.Errorf("git add: %w (stderr: %s)", err, string(stderr))
-	}
-	// Reset to the target SHA while keeping all changes staged
-	if _, stderr, err := a.runner.Run(ctx, worktreePath, "git", "reset", "--soft", targetSHA); err != nil {
-		return fmt.Errorf("git reset --soft %s: %w (stderr: %s)", shortSHA(targetSHA), err, string(stderr))
-	}
-	// Amend the target commit with the staged changes
-	if _, stderr, err := a.runner.Run(ctx, worktreePath, "git", "commit", "--amend", "--no-edit"); err != nil {
-		return fmt.Errorf("git commit --amend: %w (stderr: %s)", err, string(stderr))
-	}
-	return nil
-}
-
-// gitSquashCommits squashes all commits since the origin default branch into a single commit.
-func (a *Agent) gitSquashCommits(ctx context.Context, worktreePath string, issueNumber int, issueTitle string) error {
-	// Get all commit messages since origin default branch
-	logOut, _, err := a.runner.Run(ctx, worktreePath, "git", "log", a.originDefaultBranch()+"..HEAD", "--format=%B")
-	if err != nil {
-		return fmt.Errorf("getting commit messages: %w", err)
-	}
-	commitMessages := strings.TrimSpace(string(logOut))
-
-	// Reset to origin default branch while keeping changes staged
-	if _, stderr, err := a.runner.Run(ctx, worktreePath, "git", "reset", "--soft", a.originDefaultBranch()); err != nil {
-		return fmt.Errorf("git reset --soft: %w (stderr: %s)", err, string(stderr))
-	}
-
-	// Unstage .pr-body.md if Claude accidentally staged it — it must not be committed.
-	a.runner.Run(ctx, worktreePath, "git", "restore", "--staged", ".pr-body.md") //nolint:errcheck // best-effort
-
-	// Create a single commit with a meaningful message.
-	// Strip any Signed-off-by trailers Claude may have added to individual commits
-	// before building the body, then append exactly one canonical trailer at the end.
-	commitMsg := fmt.Sprintf("Fix #%d: %s", issueNumber, issueTitle)
-	if commitMessages != "" {
-		commitMsg += "\n\n" + stripSignedOffBy(commitMessages)
-	}
-	if a.cfg.SignedOffBy != "" {
-		commitMsg += fmt.Sprintf("\n\nSigned-off-by: %s", a.cfg.SignedOffBy)
-	}
-
-	if _, stderr, err := a.runner.Run(ctx, worktreePath, "git", "commit", "-m", commitMsg); err != nil {
-		return fmt.Errorf("git commit after squash: %w (stderr: %s)", err, string(stderr))
-	}
-
-	return nil
-}
-
-// deleteRemoteBranch removes a branch from the push remote.
-func (a *Agent) deleteRemoteBranch(ctx context.Context, worktreePath, branchName string) {
-	pushRemote := a.pushRemote()
-	_, stderr, err := a.runner.Run(ctx, worktreePath, "git", "push", pushRemote, "--delete", branchName)
-	if err != nil {
-		a.logger.Warn("failed to delete remote branch", "branch", branchName, "error", err, "stderr", string(stderr))
-	} else {
-		a.logger.Info("deleted remote branch", "branch", branchName)
-	}
-}
-
-// gitPush pushes the current branch to the push remote.
-func (a *Agent) gitPush(ctx context.Context, worktreePath string, force bool) error {
-	if a.cfg.DryRun {
-		a.logger.Info("[dry-run] would push", "worktree", worktreePath, "force", force)
-		return nil
-	}
-	pushRemote := a.pushRemote()
-
-	// Get the current branch name for the refspec
-	branchOut, _, err := a.runner.Run(ctx, worktreePath, "git", "rev-parse", "--abbrev-ref", "HEAD")
-	if err != nil {
-		return fmt.Errorf("getting branch name: %w", err)
-	}
-	branch := strings.TrimSpace(string(branchOut))
-
-	args := []string{"push", pushRemote, "HEAD:" + branch}
-	if force {
-		// Fetch first so --force-with-lease has up-to-date tracking refs
-		a.runner.Run(ctx, worktreePath, "git", "fetch", pushRemote) //nolint:errcheck // best-effort
-		args = append(args, "--force-with-lease")
-	}
-
-	if _, stderr, err := a.runner.Run(ctx, worktreePath, "git", args...); err != nil {
-		return fmt.Errorf("git push: %w (stderr: %s)", err, string(stderr))
-	}
-	return nil
-}
-
-// getPRCommits returns the list of commits in the PR (commits between origin default branch and HEAD).
-func (a *Agent) getPRCommits(ctx context.Context, worktreePath string) []Commit {
-	logOut, _, err := a.runner.Run(ctx, worktreePath, "git", "log", a.originDefaultBranch()+"..HEAD", "--format=%H %s")
-	if err != nil {
-		return nil
-	}
-
-	lines := strings.Split(strings.TrimSpace(string(logOut)), "\n")
-	commits := make([]Commit, 0, len(lines))
-	for _, line := range lines {
-		line = strings.TrimSpace(line)
-		if line == "" {
-			continue
-		}
-		parts := strings.SplitN(line, " ", 2)
-		if len(parts) < 2 {
-			continue
-		}
-		commits = append(commits, Commit{
-			SHA:     parts[0],
-			Subject: parts[1],
-		})
-	}
-	return commits
-}
-
-// hasFixupCommits returns true if there are any fixup commits in the current branch.
-func (a *Agent) hasFixupCommits(ctx context.Context, worktreePath string) bool {
-	logOut, _, err := a.runner.Run(ctx, worktreePath, "git", "log", a.originDefaultBranch()+"..HEAD", "--format=%s")
-	if err != nil {
-		return false
-	}
-	return strings.Contains(string(logOut), "fixup!")
-}
-
-// gitRebaseWithRetry attempts a rebase and, if it fails due to unstaged changes
-// (e.g. upstream file deletions), cleans the worktree and retries once.
-// Returns the stderr output (as a string) and any error from the final rebase attempt.
-func (a *Agent) gitRebaseWithRetry(ctx context.Context, worktreePath string, prNumber int) (string, error) {
-	_, stderr, rebaseErr := a.runner.Run(ctx, worktreePath, "git", "rebase", a.originDefaultBranch())
-	if rebaseErr != nil && isUnstagedChangesError(string(stderr)) {
-		// Upstream file deletions can leave unstaged changes that block rebase.
-		// Clean the worktree and retry.
-		a.logger.Info("cleaning unstaged changes before rebase retry", "pr", prNumber)
-		a.runner.Run(ctx, worktreePath, "git", "checkout", "--", ".") //nolint:errcheck // best-effort
-		_, stderr, rebaseErr = a.runner.Run(ctx, worktreePath, "git", "rebase", a.originDefaultBranch())
-	}
-	return string(stderr), rebaseErr
-}
-
-// gitAutosquashRebase performs an interactive rebase with autosquash to merge fixup commits.
-func (a *Agent) gitAutosquashRebase(ctx context.Context, worktreePath string) error {
-	// Set GIT_SEQUENCE_EDITOR to cat so rebase doesn't wait for interactive input
-	_, stderr, err := a.runner.Run(ctx, worktreePath, "sh", "-c",
-		fmt.Sprintf("GIT_SEQUENCE_EDITOR=cat git rebase -i --autosquash %s", a.originDefaultBranch()))
-	if err != nil {
-		return fmt.Errorf("git rebase --autosquash: %w (stderr: %s)", err, string(stderr))
-	}
-	return nil
-}
-
 // issueAssignedTo returns true if the given user is among the issue's assignees.
 func issueAssignedTo(issue Issue, user string) bool {
 	return slices.Contains(issue.Assignees, user)
@@ -1978,63 +453,6 @@ func isConflictError(stderr string) bool {
 // unstaged changes are blocking the rebase (e.g. upstream file deletions).
 func isUnstagedChangesError(stderr string) bool {
 	return strings.Contains(strings.ToLower(stderr), "unstaged changes")
-}
-
-// resolveConflictsParallel invokes the coding agent to resolve conflicts for a list of tasks in parallel.
-func (a *Agent) resolveConflictsParallel(ctx context.Context, tasks []conflictTask) {
-	runSequential(ctx, tasks, func(ctx context.Context, task conflictTask) {
-		// Get commit count before invoking agent and validate capture
-		commitsBefore := a.getPRCommits(ctx, task.work.WorktreePath)
-		if commitsBefore == nil {
-			a.logger.Error("failed to capture commits before resolution", "pr", task.work.PRNumber)
-			return
-		}
-		commitCountBefore := len(commitsBefore)
-
-		prompt := buildConflictResolutionPrompt(*task.work, a.originDefaultBranch(), a.cfg.SignedOffBy)
-		_, err := a.codeAgent.Run(ctx, a.runner, task.work.WorktreePath, prompt, a.logger, true)
-		if err != nil {
-			a.logger.Error("agent failed to resolve conflicts", "pr", task.work.PRNumber, "error", err)
-			// Post a hidden marker so deduplication skips this SHA on the next cycle
-			_ = a.gh.AddIssueComment(ctx, a.cfg.Owner, a.cfg.Repo, task.work.PRNumber,
-				fmt.Sprintf("<!-- oompa-bot rebase:%s -->", shortSHA(task.headSHA)))
-			return
-		}
-
-		// Verify that no unexpected new commits were created
-		commitsAfter := a.getPRCommits(ctx, task.work.WorktreePath)
-		commitCountAfter := len(commitsAfter)
-
-		if commitCountAfter > commitCountBefore {
-			a.logger.Warn("conflict resolution created new commits instead of resolving within rebase",
-				"pr", task.work.PRNumber,
-				"before", commitCountBefore,
-				"after", commitCountAfter,
-				"new_commits", commitCountAfter-commitCountBefore)
-
-			if err := a.gh.AddIssueComment(ctx, a.cfg.Owner, a.cfg.Repo, task.work.PRNumber,
-				fmt.Sprintf("⚠️ Conflict resolution created %d unexpected new commit(s) instead of resolving within the rebase flow.\n\n"+
-					"Expected: %d commits (original structure preserved)\n"+
-					"Got: %d commits (new commits added)\n\n"+
-					"Please review the commit history and manually squash or rebase to preserve the original commit structure.\n\n%s",
-					commitCountAfter-commitCountBefore, commitCountBefore, commitCountAfter, a.botComment())); err != nil {
-				a.logger.Error("failed to log warning to github", "pr", task.work.PRNumber, "error", err)
-			}
-		}
-
-		// Push the rebased branch
-		if err := a.gitPush(ctx, task.work.WorktreePath, true); err != nil {
-			a.logger.Error("failed to push after conflict resolution", "pr", task.work.PRNumber, "error", err)
-			// Post a hidden marker so deduplication skips this SHA on the next cycle
-			_ = a.gh.AddIssueComment(ctx, a.cfg.Owner, a.cfg.Repo, task.work.PRNumber,
-				fmt.Sprintf("<!-- oompa-bot rebase:%s -->", shortSHA(task.headSHA)))
-		} else if !a.ShouldSkipComment("conflict") {
-			if err := a.gh.AddIssueComment(ctx, a.cfg.Owner, a.cfg.Repo, task.work.PRNumber,
-				fmt.Sprintf("Rebased commit %s on main and pushed (conflicts resolved).\n\n%s", shortSHA(task.headSHA), a.botComment())); err != nil {
-				a.logger.Error("failed to log success to github", "pr", task.work.PRNumber, "error", err)
-			}
-		}
-	})
 }
 
 // markIssueFailed marks an issue as failed, unassigns the agent, and adds the ai-failed label.
