+++ specs/architecture.md
@@ -2,8 +2,22 @@
 
 Single long-running Go binary with a sequential polling loop.
 
-- No webhooks (avoids inbound connectivity requirements)
-- No goroutine-per-issue (keeps it simple and debuggable)
+## Design Decisions
+
+- **Polling, not webhooks**: Avoids inbound connectivity requirements, NAT traversal, and webhook secret management. The agent runs behind firewalls and on developer machines without exposing ports. The trade-off is higher latency (bounded by `--poll-interval`) which is acceptable for code maintenance tasks that tolerate minutes of delay.
+- **Sequential, not parallel**: One issue at a time prevents race conditions on shared git state (worktrees, branches). A goroutine-per-issue model would require locking around git operations and make debugging harder. The `--max-workers` flag allows opt-in parallelism when the user accepts the complexity.
+- **Interfaces over mocks**: All external dependencies (GitHub API, Claude CLI, git) are behind Go interfaces. This enables deterministic unit tests without credentials, network access, or subprocess execution. Tests verify business logic, not integration plumbing.
+- **Stateless on disk**: The agent rebuilds state from GitHub on every startup by scanning labeled issues and matching PRs. This eliminates state file corruption, migration, and backup concerns. The trade-off is a small startup cost from API calls.
+- **CLI subprocess, not SDK**: Claude is invoked via the `claude -p` CLI rather than an SDK. This decouples the agent from Claude's release cycle and authentication flow, and allows swapping to OpenCode or other CLI agents via `--agent`.
+
+## Invariants
+
+- Claude never merges -- it only creates or updates PRs. A human must approve and merge.
+- Force-pushes use `--force-with-lease` only for history-rewriting operations (rebase, squash) to maintain clean commit history.
+- Bot-posted comments are tagged with `<!-- oompa-bot -->` markers to prevent self-reply loops.
+
+## Constraints
+
 - All GitHub interaction via `github.com/google/go-github/v84` behind a `GitHubClient` interface
 - All Claude interaction via `claude -p` CLI (headless mode) using Google Vertex AI as the backend
 - Standalone `go.mod` with one external dependency: `github.com/google/go-github/v84`
@@ -34,7 +48,13 @@ pkg/agent/
   state_test.go        -- tests for state load/save
   prompt.go            -- prompt templates
   prompt_test.go       -- tests for prompt generation
-  loop.go              -- Agent struct, processNewIssues, processReviewComments, cleanupDone
+  loop.go              -- Agent struct, NewAgent, CleanupDone, BootstrapWatchedPRs, helpers
+  issues.go            -- ProcessNewIssues
+  review.go            -- ProcessReviewComments
+  ci.go                -- ProcessCIFailures
+  conflicts.go         -- ProcessConflicts, ProcessRebase, conflict resolution
+  triage.go            -- ProcessTriageJobs, periodic CI investigation
+  git_ops.go           -- git push, amend, squash, rebase helpers
   loop_test.go         -- integration tests with all interfaces mocked
 go.mod                 -- standalone module
 ```
